Update version and add Pydantic v2 support check
The package version has been updated from 0.21.0 to 0.22.0 in "__init__.py" and in "pyproject.toml" files. Given the potential compatibility issues, a check has been added and an exception is raised in "__init__.py" if Pydantic version 2 is being used, as Aiotdlib currently does not support it. Pydantic library version in the project dependency has been updated to not include version 3.