Refactor TDJson class structure and functionality
The main functions of the TDJson class are restructured for clarity and organization, class 'CoreTDJson' is introduced as a base class which is responsible for handling the loading and interaction with the TDLib binary. Child class 'TDJson' is also introduced to handle additional functionalities. Additionally, the method _get_tdjson_lib_path has been renamed to _get_bundled_tdjson_lib_path to improve clarity.